% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier_pvalue.R
\name{bezier_plot}
\alias{bezier_plot}
\title{Plot graphic to show determination of p-value according to the drawing of Bezier curve}
\usage{
bezier_plot(df)
}
\arguments{
\item{df}{as a dataframe containing p-values associated to a number (occurrence, percentage, ...)}
}
\description{
Plot graphic to show determination of p-value according to the drawing of Bezier curve
}
