% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier_pvalue.R
\name{bezier_pvalue}
\alias{bezier_pvalue}
\title{p-value detection (Bezier curve)}
\usage{
bezier_pvalue(df, cv = "no", sp = 0.5, fraction.var = 0.005)
}
\arguments{
\item{df}{a dataframe containing p-values associated to a number (occurrence, percentage, ...)}

\item{cv}{a character ("yes" or "no") to use or not the cross validation method in order to detect the best value for span parameter}

\item{sp}{an integer or a float to use as the span parameter in the loess() function}

\item{fraction.var}{keep lines if variance of p-value occurrences is greater than a fraction of total variance}
}
\description{
Determine p-value thanks to the geometric method (Bezier curve)
}
