% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidist_pvalue.R
\name{euclidist_plot}
\alias{euclidist_plot}
\title{Plot graphic to show determination of p-value according to the euclidean distance method}
\usage{
euclidist_plot(df)
}
\arguments{
\item{df}{as a dataframe containing p-values associated to a number (occurrence, percentage, ...)}
}
\description{
Plot graphic to show determination of p-value according to the euclidean distance method
}
