% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidist_pvalue.R
\name{euclidist_treatment}
\alias{euclidist_treatment}
\title{Determination of the optimal p-value
Using the euclidean distance to identify the furthest point}
\usage{
euclidist_treatment(df)
}
\arguments{
\item{df}{as a dataframe containing p-values associated to a number (occurrence, percentage, ...)}
}
\description{
Determination of the optimal p-value
Using the euclidean distance to identify the furthest point
}
