% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Dataframe preparation}
\usage{
prepare_data(data, column, nbr.pv = 20)
}
\arguments{
\item{data}{a dataframe from which we can build a dataframe usable by the PValuator}

\item{column}{the column of interest (p-values)}

\item{nbr.pv}{the minimum number of p-values}
}
\description{
Prepare a dataframe usable by the PValuator
}
