% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier_pvalue.R
\name{tangent}
\alias{tangent}
\title{Determination of the optimal p-value
Using the Bezier method to identify the curvature of a graph}
\usage{
tangent(x.tangent, smooth)
}
\arguments{
\item{x.tangent}{as a numeric point (abscissa) to determine a tangent}

\item{smooth}{as the smoothing of the curve}
}
\description{
Determination of the optimal p-value
Using the Bezier method to identify the curvature of a graph
}
