% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidist_pvalue.R
\name{euclidist_pvalue}
\alias{euclidist_pvalue}
\title{Show p-value thanks to the distance method}
\usage{
euclidist_pvalue(df)
}
\arguments{
\item{df}{as a dataframe containing p-values associated to a number (occurrence, percentage, ...)}
}
\description{
Show p-value thanks to the distance method
}
